feat: callback for popup (WIP)
