+++ widgets/general.go
@@ -44,52 +44,3 @@ func Dropdown(
 	dropdown.SetBackgroundColor(utils.COLOR_DARK_GREY)
 	return dropdown
 }
-
-type PopupKind int
-
-const (
-	POPUP_ERROR PopupKind = iota
-	POPUP_WARNING
-)
-
-// TODO: Don't make a new popup every time, make one and
-// change the attributes every time you need a "new one"
-func Popup(kind PopupKind, text string, onEsc func()) *tview.Flex {
-	var title string
-	var borderColor tcell.Color
-
-	switch kind {
-	case POPUP_ERROR:
-		title = "Error"
-		borderColor = utils.COLOR_POPUP_RED
-	case POPUP_WARNING:
-		title = "Warning"
-		borderColor = utils.COLOR_POPUP_YELLOW
-	}
-
-	content := tview.NewTextView()
-	content.SetTitle(title)
-	content.SetText(text)
-	content.SetBorder(true)
-	content.SetBorderColor(borderColor)
-	content.SetBackgroundColor(utils.COLOR_DARK_GREY)
-
-	popup := tview.NewFlex().
-		AddItem(nil, 0, 1, false).
-		AddItem(tview.NewFlex().SetDirection(tview.FlexRow).
-			AddItem(nil, 0, 1, false).
-			AddItem(content, 10, 1, true).
-			AddItem(nil, 0, 1, false), 40, 1, true).
-		AddItem(nil, 0, 1, false)
-
-	popup.SetInputCapture(func(event *tcell.EventKey) *tcell.EventKey {
-		pressedKey := event.Key()
-		switch pressedKey {
-		case tcell.KeyESC:
-			onEsc()
-		}
-		return event
-	})
-
-	return popup
-}
